+++ buildSrc/src/main/java/Config.kt
@@ -10,7 +10,7 @@ object Config {
     val kotlinCompatibleLanguageVersion = "1.4"
 
     object BuildPlugins {
-        val androidGradle = "com.android.tools.build:gradle:7.0.4"
+        val androidGradle = "com.android.tools.build:gradle:7.1.0"
         val kotlinGradlePlugin = "gradle-plugin"
         val buildConfig = "com.github.gmazzo.buildconfig"
         val buildConfigVersion = "3.0.3"